add external database support in operator
currently, in the syndesis operator, the creation of the in-cluster
postgres instance is hardcoded. there is no way to specify an
external database through the custom resource.

this change adds a new field to the syndesis custom resource named
externalDbSecretRef which should be set to the name of a secret in
the same namespace. this secret contains information about an
external database.

this new field is optional, if it is not specified, an in-cluster
postgres will be deployed as normal, to keep the operator backwards
compatible.