refactor: cleanup PublicApiHandlerTest
`PublicApiHandlerTest` much like the whole public API is a convoluted
mess. At least having the test cleaned could help with refactoring it to
a smaller mess.

Several issues with the test:
 - test mostly exercised mocks
 - assertions tested to broadly (i.e. _some_ integration was updated)
 - spaghetti mock usage, where by any change to a mock configuration
   would introduce issues in several test cases
 - no exception assertions were used

This refactors the test to use AssertJ instead of JUnit Hamcrest
assertions which are much more versatile; cleans up mock usage and does
general refactoring for code quality.