refactor: proper exception assertion
Annotating tests with `@Test(expected = SomeException.class)` does poor
assertion of the exception that's expected to be thrown and possibly
fails the assertion as the exception can be thrown from any line in the
test method and with any arbitrary message or cause.

This construct should be replaced with AssertJ's
`assertThatExceptionOfType` et al. exception assertions[1].

[1] https://joel-costigliola.github.io/assertj/assertj-core-features-highlight.html#exception-assertion