fix(release): don't perform Java release for UI
This makes sure that the `non-ui-related-so-disabled` profile is active
in cases where other profiles are activated via command line, as is the
case in release when we specify `-Prelease`.

Unfortunately `<activeByDefault>` activates the profile only in the case
other profiles are not activated[1].

Also ensures additional plugin executions, not related to the UI build
are not performed by adding them to the non-phase.

[1] https://issues.apache.org/jira/browse/MNG-4917