fix(release): dist attachment and POM signature
We do not wish to publish the `ui-react-x.y.z-dist.tar.gz` artefact to
Maven central, so we need to disable the attachment of artifact from the
Maven reactor. Being that the ui-react module is the last module in the
reactor and that staging of artefacts will happen on the last module
there is no way to escape staging of the POM file of the ui-react. So
for the staged repository to be released we can't skip the signing of
the POM file.

So this will in fact publish the ui-react module to Maven central, but
hopefully only of the POM.

[ci skip]