chore: add `--nodev` option to README
We already point to the developer handbook. The quickstart here should be targeted towards non-developers. If we don't have the `--nodev` option on the CLI then the image streams will point to local OpenShift registry that won't have any images until we build and push to that registry via S2I.