fix(install): platform and binary name
We renamed the operator binary from `operator` to `syndesis` to
`syndesis-operator` (I think that was how the lineage went). The current
binary is named `syndesis-operator`, so we need to use that name as it's
the filename that will be created when we untar the release from GitHub.

The linux platform tar archive has `linux-amd64` in it, we need the to
reflect that.