fix(c/rest-swagger): array query parameters
This adds support for setting multiple values for query parameters in
the custom API client connectors.

For example query parameters that are of array type and for which
multiple values are given now appear multiple times as query parameter
each time with one of the values.

For example:

    /v2/pet/findByStatus?status=available&status=pending

For `status` having a values `available` and `pending`.

Ref https://issues.jboss.org/browse/ENTESB-12139