chore(cli): add option to reset everything
This adds `--nuke` option to the `dev --cleanup` command:

    syndesis dev --cleanup --nuke

When specified tables in the syndesis database with user data will be
truncated and any causal user created OpenShift object will be deleted;
effectively reseting the environment back to the initial state.

In the process the syndesis-server pod will be restarted, this is to
load the syndesis database with the initial data.