refactor(mongodb): database directory and output
This creates the database in a temporary directory instead of $HOME
making the tests repeatable and allow them to run concurrently and in
environments where $HOME is not writable (CI).

Redirects the output to `target/test.log` to speed up the build and keep
the output separate from the console output of the build.