chore(it): remove stub www.googleapis.com server
This removes the test and the stub server configuration from the Google
Calendar connector integration test. As noted previously it's really
difficult to get this to work.

The test still has value as the issue the test tries to test manifested
itself on the integration startup and the test currently tests that the
integration successfully starts.