fix(c/google-calendar): revert dependency upgrade
Using a newer dependency (v3-rev20190929-1.30.3 vs v3-rev291-1.22.0)
lead to a runtime issue due to a change in the method signature. This
reverts to the `com.google.apis:google-api-services-calendar` version
used by Camel 2.21.

This also attempts to add an integration test (webhook-to-create-event)
and fails miserably at that. The endpoint for the API hosted at
`www.googleapis.com` is not configurable in the google-calendar
component and one possible solution for that would be to use a local
stub instead. For that an entry to the `/etc/hosts` file was added to
resolve `www.googleapis.com` as `127.0.0.1` and the port of the stub
server running on the host was forwarded to the integration container
network. This ended up with the forwarded port accessible in the port
forwarding container and not in the integration container. The IP of the
port forwarding container is not (without reflection) accessible at the
time on configuring the addition to the `/etc/hosts` so using it this
way did not help. Also the stub Jetty server could not be configured to
run with http/TLS which would be requirement to run it as the
`www.googleapis.com` is accessed over HTTPS. The Jetty stub server
failed the handshake by responding as:

```
read from 0x5648ed3d0360 [0x5648ed3def33] (5 bytes => 5 (0x5))
0000 - 15 03 03 00 02                                    .....
read from 0x5648ed3d0360 [0x5648ed3def38] (2 bytes => 2 (0x2))
0000 - 02 28                                             .(
140600789931840:error:14094410:SSL routines:ssl3_read_bytes:sslv3 alert handshake failure:ssl/record/rec_layer_s3.c:1543:SSL alert number 40
```

Possibly due to the compression filter.