fix(build): s2i build within a container
`maven.repo.local` is a legacy property there for compatibility with
Maven 1, and the replacement is `settings.localRepository. When running
the build in a container the `user.home` Java property was set correctly
but the property `maven.repo.local`, was resulting into `null` (checked
via `mvn help:evaluate`), while `settings.localRepository` evaulated to
correct path of the Maven local repository.

Not sure what circumstances necessitated the workaround in the POM of
the S2I module, this seems to be solving the problem more _correctly_.