refactor(s2i): replace shell script with Maven
The goal of the S2I build is to generate a base S2I image that will
contain the already prepopulated (cached) Maven repository to eliminate
the need to download dependency artifact download when building the
integration pod.

It does this by using the project generator to generate a project with
dependencies on all connectors and the integration runtime. Next that
project is built using Maven to gather all the dependencies.

In the evolution of the build of this module we ended up executing the
generated project twice and somehow forgot what the goal of the S2I
build is: to gather the dependencies.

This replaces the invocation of the shell script with a series of Maven
plugins that will:
 - run the server-builder-image-generator to generate the Maven project
   with all connectors as dependencies
 - run the Maven build of the generated Maven project with the goal
   `dependency:go-offline` to gather all the dependencies in a
   repository within the target directory
 - run the fabric8 Maven plugin to build the syndesis-s2i image