refactor(s2i): two step Maven invocation
This refactors the s2i build further to include a two step Maven
invocation: now the dependencies of the generated project are gathered
twice, first using the merged local and image settings and caching the
result in the local Maven repository and secondly gathering the
dependencies from the local Maven repository only to the repository
directory used to generate the image.

This way the dependencies are cached in the local repository and are not
fetched in subsequent invocations. Given no changes clean build takes
around ~1,5min (on my machine), non-clean build takes ~25sec (on my
machine).