chore(build): setup maven-buildtime-extension
Seems that we were missing the setup for the maven-buildtime-extension.
The intent was there to configure, the configuration wasn't applied as
the `.mvn` directory was at the wrong level and the property needed to
configure it was missing.

Now timing information for each plugin/module will be printed to stdout.