chore: parameterize PostgreSQL version
This introduces test resource filtering so that the `generated-pom.xml`
can be filtered to include the PostgreSQL version dictated by the
`pom.xml`.

Next it refactors the test so that AssertJ `isXmlEqualTo` is used, which
disregards insignificant formatting issues when comparing XML documents.

In order for the `isXmlEqualTo` to function properly, XML API and Xerces
Xalan dependencies were excluded. It's a bit of a mistery why those
would be included their inclusion would trump the XML API/Xerces/Xalan
provided with the JDK which would lead to all sorts of compatibility
issues.

# Conflicts:
#	app/dv/src/test/resources/generated-pom.xml