fix(openshift): cascade `BuildConfig` deletion
When deleting an integration the `Build` would be left over preventing
the future integration with the same name, and hence the same `Build`
name to be built and started. This makes sure that the deletion of
`BuildConfig` is propagated explicitly by using `Foreground`[1]
cascading deletion.

Ref. https://issues.redhat.com/browse/ENTESB-12493

[1] https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/#foreground-cascading-deletion