chore(build): make buildtime extension opt-in
The buildtime Maven extension generates a long list of plugin execution
times at the end ov Maven execution. When an error is reported in the
build it's followed by this long list of build plugin timings. This
makes it cumbersome to scroll to the error.

This makes the buildtime extension opt-in, i.e. to activate the
extension `-Dbuildtime.output.log=true` must be specified.