fix: connector properties fetching
This fixes and refactors connector properties fetching from the meta
pod. I've found the implementation to be overly optimistic and incorrect
in assumptions. As it was, any error from the meta pod would be
converted to a HTTP successful response and no error handling would be
performed. This would lead to issues when the meta pod would respond
with an error and the an attempt to parse the response to the expected
format would result in errors like `NullPointerException`.

Also adds the missing tests.

Fixes #7504