fix(ui): `value` instead of `label` in typeahead
patternfly-react doesn't seem to allow us to have one value displayed
and another value bound/kept as value. That means that our `enum`
`label/value` can't be used as we intended: to have the `label`
displayed and to have the `value` kept as `configuredProperty`. So this
simply displays the `value` from the `enum` instead, which will make
that value bound to `configuredProperty` as well.