chore: update swagger-core
To help with #7556 this updates the `swagger-core` to the latest version
(2.1.1).

This will change the version of the served API documentation from
Swagger 2.0 to OpenAPI 3.0.

For Syndesis Meta I'm not aware of any client using this API
documentation so this should be safe to do.

For the Syndesis Server more extensive changes needed to be done:

 * annotations used needed to change
 * tooling around generating Asciidoc/HTML of the API needed to be
   upgraded / changed
 * endpoints changed from `/api/v1/[internal/]swagger.[yaml|json]` to
   `/api/v1/[internal/]openapi.[yaml|json]`

Changes to the UI build need to be done, as the current tool chain
relies on Swagger 2.0 file.