fix: server/meta startup
This is a dirty fix to circumvent "clever" classpath scanning from
`swagger-jaxrs` that fails on Reflections tripping over closing a
directory in a JAR file which in turn closes the whole JAR file stream
and prevents the further scanning and fails the app startup.

A proper fix is upgrading to the 2.x line of `swagger-core` which
doesn't rely on Reflections to do classpath scanning.

Ref #7556