fix(c/sql): use standard column names
When fetching the list and input/output parameters of stored procedures
we're relying on `java.sql.DatabaseMetaData`.

For PostgreSQL database instead of the `DatabaseMetaData::getProcedures`
we were using `DatabaseMetaData::getFunctions`, this, by standard should
return `FUNCTION_NAME` and `FUNCTION_TYPE` columns. For other databases
we're using `DatabaseMetaData::getProcedures` returning `PROCEDURE_NAME`
and `PROCEDURE_TYPE` columns.

For all databases we're trying to read `PROCEDURE_NAME` and
`PROCEDURE_TYPE` columns, regardless of the `ResultSet`'s origin:
`DatabaseMetaData::getFunctions` or `DatabaseMetaData::getProcedures`.

The issue reported was caused by the missing `PROCEDURE_NAME` column for
the PostgreSQL database. So seemingly the
`DatabaseMetaData::getFunctions` for PostgreSQL was returning
`PROCEDURE_NAME` column previously.

The database driver version for PostgreSQL did not change, so that
leaves that there was a change in PostgreSQL server that made the
`DatabaseMetaData::getFunctions` return the standard `FUNCTION_NAME` /
`FUNCTION_TYPE` columns vs what we were seeing previously
(`PROCEDURE_NAME` and `PROCEDURE_TYPE`).

This change switches to use the `DatabaseMetaData::getProcedures` method
for PostgreSQL database as well, with the caveat that this might fail on
older versions of the PostgreSQL database.

Ref. https://issues.redhat.com/browse/ENTESB-12415