fix(tools): fix pre-commit configuration
Seems that in the Husky upgrade the hooks interact with the shell, and
that the shell expansion is now in place. That is, running the
pre-commit hook would fail with an error like:

    "zsh:1: no matches found: pre-commit:*"

This simply adds quotes around the hooks so the arguments for
`npm-run-all` are taken verbatim and not expanded by the shell.

Also it seems that the `git add` is no longer required so this removes
it.