fix(model): openapi schema generation
The current state of `ModelResolver`/`ModelConverter` in Swagger Core
tests for the existence of `JsonIgnore` property, not for the value it
holds, so `@JsonIgnore(false)` annotated members will be removed from
the generated JSON Schema.

This adds a custom ModelConverter that will include the member based on
the value of the `@JsonIgnore` annotation.

The Swagger Core Schema generation also detects `withId` as a separate
property and `@JsonIgnore` was added to remove it from the generated
Schema.

(cherry picked from commit 8cbd85172edd889c39b227d39e9fe97e1b07ed7a)

# Conflicts:
#	app/common/model/pom.xml