refactor(deps): remove dependency on glassfish-el
Immutables detects validation API and bakes its usage in the
`Immutable*` class in a static field. When the validation API
implementation is bootstrapped by default implementation of EL language
is also bootstrapped. This leads to a runtime dependency on
glassfish-el.

Hibernate Validator can be configured to use a
`ParameterMessageInterpolator` instead of default `MessageInterpolator`
that relies/supports on EL implementation/language.

This means that the validation messages can't use EL expressions, which
we're not using at the moment.

I was expecting that the `${...}` expressions would not work after this
change, as supposedly the `ParameterMessageInterpolator` supports only
`{...}` placeholders. This proved to be a wrong assumption and in this
was validated in the integration tests we have in the `server-endpoint`.