refactor(extension): disable repackage test
This disables the `repackage` integration test from running by default.
It takes inordinate amount of time to run due to the tendency to
download exuberant amount of Maven dependencies.

This test was added to test the Maven 3.6.3 upgrade, it could be made to
pass locally with enough patience. When the `repackage` goal was
refactored not to use Shrinkwrap and this test was added to compare the
two implementations at that point the performance characteristics were
in favor of the Maven Resolver implementation. This is a purely
subjective conclusion, since more thorough A/B tests were not performed.