chore: update RESTEasy to 4.4.2
This should only upgrade the RESTEasy to a newer version, it indeed does
that but it also moves to using Jakarta EE dependencies instead of
`javax` or `com.sun` ones. Because of many dependencies we pull in also
use the `javax` or `com.sun` dependencies this also tries to conform
those to the Jakarta EE ones by excluding the `javax` or `com.sun` ones
from the dependency tree.

This is a fairly tech debt inducing, as the exclusions we now have don't
have to carry on to the next versions, so say we update Citrus to a new
version (just to pick one) and it solely depends on Jakarta EE and
Glassfish (for JAXB implementation) dependencies, the exclusions put in
place now will become irrelevant, but without any warning/error in the
future. So might be in danger of cargo culting those exclusions even
though they are irrelevant in the future.

There is a risk of having different dependencies in meta vs at runtime
in the integration, I consider this a smaller risk as the javax
specifications are very much backward compatible.

Enforcement plugin rules were added to make sure we don't introduce the
`javax` or `org.jboss.spec` dependencies. These were also disabled for
connectors, since we do wish those there for when connector is running
in the integration, the Camel and Spring Boot BOMs should dictate the
dependencies there.