chore: Fix weird testcontainer behavior to not wait for other containers to start
Nightly integration tests were failing for quite some time because of a Http request timeout in the Http-to-AMQ tests. I boiled this down to the integration runtime container wait strategy not being able to wait for the AMQ testcontainer to be pulled and started. Therefore the test logic started way too early (while container images still being pulled) and a Http request timeout causes the tests to fail.

Now using a static initializer for AMQ testcintainer instead of the @ClassRule