fix(deps): exclude javax.annotation-api
Seems that there are duplicate but different classes in
`javax.annotation:javax.annotation-api:1.3.2` vs
`org.jboss.spec.javax.annotation:jboss-annotations-api_1.2_spec:1.0.2.Fi
nal`.

This excludes the `javax.annotation:javax.annotation-api` as the JBoss
one seems to be used more.

Caveat, on `master` we have excluded both the javax ones and the JBoss
ones in favor of the Jakarta versions.