fix(c/fhir): update FHIR connector to Camel 3.1
Seems that the default payload encoding in newer FHIR version switched
to JSON failing a number of integration tests in the FHIR connector.

To fix that I've added a `encoding=XML` configured property to the
connector definition. To prevent existing integrations using the FHIR
connector that property needs to be added to the connector definition
embedded within the flow -> action -> connection. For that a migration
was added to set it.