feat(operator): determine PostgreSQL from image
This adds logic to determine the version of the PostgreSQL from Docker
image metadata. Usually the metadata contains environment variable
`POSTGRESQL_VERSION` (e.g. `POSTGRESQL_VERSION=9.5`); I've also noticed
that there could be a `version` label also present, so that can be used
as a fallback if `POSTGRESQL_VERSION` environment variable is not
present. Caveat with using the `version` label is that some images have
the value set to `1`, go figure.

The logic performs as follows:
 1. image name (pull spec) is taken from the `DeploymentConfig`'s
    container template
 2. image name is parsed to extract the ID (sha hash)
 3. `Image` is fetched by the extracted ID
 4. from the `Image` data Docker image metadata is parsed
 5. `POSTGRESQL_VERSION` environment variable is looked up in the Docker
    image metadata, failing that
 6. `version` label is looked up in the Docker image metadata, failing
     that

The logic to determine if the database upgrade is needed now considers
this the above determined version and the constant
`wantedPostgreSQLVersion` corresponding to the version that the operator
would install.

Determining the version from just the image name, lacking the complete
pull spec (e.g. no ID) would mean contacting the image registry to
resolve the full pull spec and fetch the ID needed to fetch the `Image`.
This was discarded as an option for it would introduce additional
complexity and in it's place simple constant was added.

Consequence of that is that the `wantedPostgreSQLVersion` needs to be
updated with each upgrade of the PostgreSQL database.