fix(operator): install cli command
Not exactly sure if this is a correct solution, we need to determine the
database image and when the operator is used on the command line.

This tries to parse the configuration to determine the image of the
database in order to be able to pass it to the operator template.

This might break non-cli usage, not sure.