chore(operator): determine version of postgresql
In upgrades we need to know if we should set the `POSTGRESQL_UPGRADE`
environment variable based on the currently running (source) and the
to-be (target) versions of PostgreSQL.

This determines the version of the source PostgreSQL database by
connecting to the database and executing `SHOW server_version`. Since
we're interested only in major and minor version (e.g. 9.6, not 9.6.12)
parses the version and reduces it to a `float64` containing the version.

To determine the target database version the `syndesis-operator` pod as
defined using the OpenShift `DeploymentConfig` now contains an init
container running from the same image configured for the `syndesis-db`
database. This init container runs `postgres -V` which outputs to the
log the version. The operator parses the log output of the init
container to determine the version, also reduced to `float64` containing
only major and minor version.