fix(operator): provide default database image
When the operator is run as CLI there is no `/conf/config.yml` to be
read, as it is only put in the container image. This leads to:

```
Error: open /conf/config.yaml: no such file or directory
```

This reads the `install/operator/build/conf/config.yaml` fetches the
database image and creates a `install_defaults.go` file at build time
containing the `defaultDatabaseImage` constant with the value of the
read database image.

At runtime, if a config file was specified via `--config` parameter the
database image is read from that file and the default is replaced.

Fixes #8009