operator: Move resources cleanup from install to upgrade (#8011)
* fix: Move resources cleanup from install to upgrade

When an upgrade is kicked there are resources that needs to be deleted otherwise the operator continues updating them. This should be a step in the upgrade package rather than a hack in the install action. We have now another step (cleanup) doing this task

Dont perform scale.up step, it makes no sense since the PostUpgradeRun phase will scale up the DeploymentConfigs.

Remove todo DeploymentConfig and BuildConfig as part of the cleanup step in upgrade

Remove invalid trigger from todo BuildConfig

* fix: make it possible to change config file using flags

Now that we need the config file to install the syndesis app, we also need a way to change it using flags

* chore: Includes generated assets

* chore: Includes changes for suggestions