fix(operator): add `--operator-config` to install
This adds `--operator-config` option to the `install` subcommand of the
operator CLI, the option was previously only used for the
`install forge` subcommand, now it is hoisted to the `install` to have
the ability to specify a custom configuration for the operator if
needed.

Note that it should not be necessary to specify `--operator-config` for
any command other than the `install forge`, currently by default the
configuration setting that needs to be read from the operator
configuration is baked in the binary at build time (the
`defaultDatabaseImage`).

(cherry picked from commit 0d3a25a37c17eeb6a1990162928cde0d78bfdc7f)