fix(server): `required` when `null` raises NPE
For OpenAPI 3 documents we need to be careful about accessing the
`required` field of parameters as it might be `null`. This make sure
that `NullPointerException` will not be raised in case the value of
`required` might be `null`.

Fixes #8038

(cherry picked from commit c38b70e6791a9928248fdb94045e323d4fbe0893)