Allowing for the absence of the pg_up metric
The FuseOnlineDatabaseInstanceDown alert doesn't trigger if syndesis-db is down.
However, it does look like the `pg_up{job="syndesis-db"}` metric is no longer being scraped.
The problem is how prometheus does the `sum`.
If it tries to sum an empty result (no data), it also returns an empty result.
Therefore, we need to add a conditional for the metric being absent and alert in that scenario.