fix(server): add new micrometer tags
We're using a weird combination of Spring/Spring Boot and JAX-RS, and
any changes in Spring/Spring Boot metrics we need to replicate in JAX-RS
world. This causes an issue, like when a metric tag is not provided from
JAX-RS.

This adds the two tags that were reported as missing. For the
`exception` tag, JAX-RS doesn't offer any way of accessing it from a
filter so no way of setting that value.

Fixes #8056