fix(build): git style remote in release command
Seems that on CI we're using git remote URLs not HTTP remote URLS, this
tries to support that. Should fix the snapshot releases that fail with:

```
curl: (6) Could not resolve host: origin; Unknown error
ERROR: Cannot create release on remote github repository. Check if a release with the same tag already exists.
```

Fixes #8166