feat: add configuration field to swap between tag and sha references
Include a configuration field to allow users to chose whether docker images are referenced by tag or by sha. This change is required for Openshift 4.5+ OLM based installation in order to allow for disconnected(offline) installations.

Whether to use SHA or Tags can be set in the syndesis operator configuration file. By default, Tag is enabled. This change doesn't require any special treatment for upstream, but for downstream some `sed` magic is required to change the configuration file and enable SHA references.

At this point, it is not possible to enable it via syndesis api (custom resource). This can be added later, I am just not sure we want to enable this feature via the syndesis api.

BREAKING CHANGE: All references to Syndesis.Components.$COMPONENT.Image and Syndesis.Addons.$ADDON.Image(and until this point this was a reference to the docker image + tag) are now changed to Syndesis.Components.$COMPONENT.Tag and Syndesis.Addons.$ADDON.Tag respectively, since the name "Image" was ambiguous. There is an extra configuration field Syndesis.Components.$COMPONENT.SHA and Syndesis.Addons.$ADDON.SHA to refer to the sha version of the docker image for any component or addon.

Signed-off-by: Luis Garcia Acosta <lgarciaac@gmail.com>