chore: Fix api client versions used in Google connectors
Revert dependabot update of api client versions. This leads to unexpected behavior and MethodNotFound errors because underlying Google Camel components still use 1.22.0 as client and api versions. If we want to update to 1.34.0 we need to update Camel components first. Otherwise we end up having mixed versions on the classpath leading to errors. Google sheets integration tests were failing because of that.