chore(build): remove Maven repositories
With multiple repositories defined Maven needs to contact more than a
single Maven repository, in worst case all of them, to download the
dependency. This slows down the build on non-cached repositories and, on
occasions, fails the build due to the repository being inaccessible.

This removes additional repository declarations, leaving only the ones
that are needed for the build to succeed. Namely, jitpack repository for
spring-social fork we rely on, and atlassian repository for the JIRA
client.

This reduces the build time for a `flash` build from 10:18min to 6:53min
or roughly by a third.