fix(test): testcontainers on CI
This should fix issues with running testcontainers on CI where SELinux
is enabled. This solves the issue by adding the files of the integration
to the image at image build time instead of mounting the volume at
container runtime. When adding the files as a part of the build we can
control permissions whereas when mounting a volume we need to make sure
that the SELinux context is preserved and that the permissions on files
are sufficient enough taking into account the permissions of the user
running the container.

Also configures that the Ryuk container runs as privileged, this is
required so it can access the Docker socket mounted from host in
environments where SELinux is enabled.