refactor(operator): database upgrade
This moves the database upgrade into the operator upgrade life cycle.
Now the database upgrade will only be attempted if the version of
Syndesis changes and the operator kicks of the upgrade.

The upgrade is performed by creating a Deployment of the PostgreSQL
target image with the `POSTGRESQL_UPGRADE=copy` environment variable.
Presence of this variable triggers the execution of `pg_upgrade` in the
container startup. After the deployment scales up to one replica,
meaning that the database upgrade has completed, as the container is now
ready, the Deployment is removed, shutting down the container.

Upgrade needs to be done after migrations, this is less optimal, the
migrations need a running database, but the database needs to be scaled
down to be upgraded. If the database is not scaled down during upgrade
the database instance spun up for the upgrade can't obtain the lock on
the database and will fail. After upgrade the database can't be scaled
up again as the data files have been migrated to the new version and the
database deployment points to the old version image. Running the old
database version (image) with new data files will fail.