fix(test): copy project jar race condition
The S2I test container is considered started when the output of the
container shows `... done`, that also coincides with the
`/usr/local/s2i/assemble` script finishing and the process stopping,
stopping the container with it. Using
`GenericContainer::copyFileFromContainer` on a stopped container fails
with `IllegalStateException: copyFileFromContainer can only be used when the Container is created`
so the container needs to be running at the point we reach to copy the
files outside of it.

This adds a `&& sleep infinity`, so that after the `assemble` script
finishes the `sh` process in the container remains running so we can
guarantee that `copyFileFromContainer` can be performed afterwards.

The container is stopped when test finishies and/or ryuk garbage
collects spawned containers.

(cherry picked from commit e1fa83333833a2b1ad3ff3612af08e5e56088a8e)