chore: Change env variable names for replacing operand images
With CLOUDBLD-26 at sight, it is a requirement that operand images are replaceable using environment variables, and such environment variables must follow a name convention (RELATED_IMAGE_*). We have in place the mechanisms to overwrite operand images.

This commit changes the name of the environment variables to meet CLOUDBLD-26 requirements.

For sure, some adjustments in QE are needed

Signed-off-by: Luis Garcia Acosta <lgarciaac@gmail.com>