fix(#8855): Fix application.properties for Camel K runtime
Camel K runtime is more restrictive when it comes to settings and properties that are set but do not match the provided application components. When a property (e.g. camel.component.servlet.header-filter-strategy) is set but not used by any active component in the application Camel K runtime will raise errors. Therefore we need to add servlet mapping component configuration only when a service is exposed in the integration.