chore: Remove CircleCi nightly job
Nightly job that runs the integration tests is broken for quite some time due to the update to JDK 11. First analysis gave the impression that the job is not easy to fix because of the build executor being an ubuntu virtual machine that only supports Java 1.8 at the moment. As we also run the integration tests for each PR using github actions the effort to also fix the nightly on CircleCi might not be worth it.