fix: align server and pg_dump versions
We're using `crunchydata/crunchy-pgdump` for the backup and restore of
the PostgreSQL database. We need to make sure that the versions there
align with the versions of PostgreSQL server used. Older versions of
`pg_dump` will not dump newer versions of server.

We're using `centos/postgresql-10-centos7` which is currently at version
10.12, this uses `crunchydata/crunchy-pgdump:centos7-10.11-4.2.1` which
is at 10.11. For some reason version
`crunchydata/crunchy-pgdump:centos7-10.12-4.1.2` contains version 12.2.

So we can't use exact (10.11) version for both server and `pg_dump`.