Update dependencies ENTESB-10454
This is related to "ENTESB-10454 Syndesis alignment problems"
https://issues.redhat.com/browse/ENTESB-10454

Changes to maven-plugin, the AbstractMojo from spring-boot-maven-plugin already provides a
project attribute, but once the plugin runs, the project attribute is
not set. So, renaming just works.

Chnages to jakarta artifacts are related to springboot upgrade,
the artifact javax.el is deprecated.