fix(operator): panic on go.mod go.sum search
The `init` function in `pkg/build/pkg.go` relies on having the project
sources checked out by tring to locate `go.mod` or `go.sum` files. These
are not present when the binary is run outside of the source tree.

The lookup is used to determine the correct path to the resource asset
files.

We can rely on the behaviour of golang tooling to set the working
directory when build/code generation is run, and use relative paths
instead. This eliminates the need to have the `init` function.